requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.38
Werkzeug==3.1.4
zipp==3.23.0
ollama==0.6.1
